refactor(examples): drop fork-test-covered personas, remove shim, add no-anvil snippets
